+++ src/main/java/com/javampire/openscad/action/SetModelColorAction.java
@@ -1,12 +1,14 @@
 package com.javampire.openscad.action;
 
-import com.intellij.openapi.actionSystem.ActionPlaces;
 import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.actionSystem.Presentation;
+import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.IconLoader;
+import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.ColorPicker;
+import com.javampire.openscad.editor.OpenSCADPreviewFileEditor;
 import com.javampire.openscad.editor.OpenSCADPreviewFileEditorConfiguration;
-import org.cef.browser.CefBrowser;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -16,22 +18,28 @@ public class SetModelColorAction extends OpenSCADAction {
 
     @Override
     public void update(@NotNull final AnActionEvent event) {
-        super.update(event);
-        if (ActionPlaces.EDITOR_TOOLBAR.equals(event.getPlace())) {
-            final Presentation presentation = event.getPresentation();
-            if (presentation.isEnabled()) {
-                presentation.setText("Set Model Color");
-                presentation.setDescription("Set preview model color");
-                presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/colorPicker.svg", getClass()));
-            }
+        final Presentation presentation = checkOpenSCADPrerequisites(event);
+        if (presentation.isVisible()) {
+            presentation.setText("Set Model Color");
+            presentation.setDescription("Set preview model color");
+            presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/colorPicker.svg", getClass()));
         }
     }
 
     @Override
     public void actionPerformed(@NotNull AnActionEvent event) {
-        final OpenSCADPreviewFileEditorConfiguration editorConfig = event.getData(OpenSCADDataKeys.EDITOR_CONFIG);
-        final CefBrowser browser = event.getData(OpenSCADDataKeys.PREVIEW_BROWSER);
-        if (editorConfig != null && browser != null) {
+        OpenSCADPreviewFileEditor previewFileEditor = event.getData(OpenSCADDataKeys.PREVIEW_EDITOR);
+
+        if (previewFileEditor == null) {
+            final Project project = event.getProject();
+            final VirtualFile scadFile = event.getData(CommonDataKeys.VIRTUAL_FILE);
+            if (project != null && scadFile != null) {
+                previewFileEditor = getOpenSCADPreviewFileEditor(project, scadFile);
+            }
+        }
+
+        if (previewFileEditor != null) {
+            final OpenSCADPreviewFileEditorConfiguration editorConfig = previewFileEditor.getEditorConfig();
             ColorPicker.showColorPickerPopup(
                     event.getProject(),
                     editorConfig.getModelColor(),