+++ src/main/java/com/javampire/openscad/action/ToggleGridAction.java
@@ -1,10 +1,9 @@
 package com.javampire.openscad.action;
 
-import com.intellij.openapi.actionSystem.ActionPlaces;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.Presentation;
 import com.intellij.openapi.util.IconLoader;
-import com.javampire.openscad.editor.OpenSCADPreviewFileEditorConfiguration;
+import com.javampire.openscad.editor.OpenSCADPreviewFileEditor;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -14,27 +13,24 @@ public class ToggleGridAction extends OpenSCADAction {
 
     @Override
     public void update(@NotNull final AnActionEvent event) {
-        super.update(event);
-        if (ActionPlaces.EDITOR_TOOLBAR.equals(event.getPlace())) {
-            final Presentation presentation = event.getPresentation();
-            if (presentation.isEnabled()) {
-                presentation.setText("Toggle Grid");
-                presentation.setDescription("Show or hide the preview grid");
-                final OpenSCADPreviewFileEditorConfiguration editorConfig = event.getData(OpenSCADDataKeys.EDITOR_CONFIG);
-                if (editorConfig != null && Boolean.TRUE.equals(editorConfig.getShowGrid())) {
-                    presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/grid.svg", getClass()));
-                } else {
-                    presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/grid_grayed.svg", getClass()));
-                }
+        final Presentation presentation = checkOpenSCADPrerequisites(event);
+        if (presentation.isVisible()) {
+            presentation.setText("Toggle Grid");
+            presentation.setDescription("Show or hide the preview grid");
+            final OpenSCADPreviewFileEditor previewEditor = event.getData(OpenSCADDataKeys.PREVIEW_EDITOR);
+            if (previewEditor != null && Boolean.TRUE.equals(previewEditor.getEditorConfig().getShowAxis())) {
+                presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/grid.svg", getClass()));
+            } else {
+                presentation.setIcon(IconLoader.getIcon("/com/javampire/openscad/icons/grid_grayed.svg", getClass()));
             }
         }
     }
 
     @Override
-    public void actionPerformed(@NotNull AnActionEvent event) {
-        final OpenSCADPreviewFileEditorConfiguration editorConfig = event.getData(OpenSCADDataKeys.EDITOR_CONFIG);
-        if (editorConfig != null) {
-            editorConfig.toggleShowGrid();
+    public void actionPerformed(@NotNull final AnActionEvent event) {
+        final OpenSCADPreviewFileEditor previewEditor = event.getData(OpenSCADDataKeys.PREVIEW_EDITOR);
+        if (previewEditor != null) {
+            previewEditor.getEditorConfig().toggleShowGrid();
         }
     }
 }