+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 - [Issue-80](https://github.com/ldenisey/idea-openscad/pull/80) Adding formatting of unary operators, thanks to [Bert Baron](https://github.com/bertbaron)
+- [Issue-81](https://github.com/ldenisey/idea-openscad/issues/81) Keyboard shortcut to refresh preview
+- [Issue-99](https://github.com/ldenisey/idea-openscad/issues/99) Auto-refresh preview on file save
 
 ### Changed
 - [Issue-83](https://github.com/ldenisey/idea-openscad/issues/83) Maintain viewport on preview refresh
@@ -13,7 +15,6 @@
 - [Issue-86](https://github.com/ldenisey/idea-openscad/pull/86) Fix EAP incompatibilities : removal of third party library
 - Unifying settings code example
 - [Issue-98](https://github.com/ldenisey/idea-openscad/pull/98) Migrate JS script to vanilla THREE.js
-- [Issue-81](https://github.com/ldenisey/idea-openscad/issues/81) Keyboard shortcut to refresh preview
 
 ### Removed